Move conftest for integration tests
